Code improvement for intermediate pickPhysicalDevice
In one of the early, intermediate representations of isDeviceSuitable
the code uses the class variable 'physicalDevice' instead of the
argument passed in, 'device'.

Even though this code is iterated on, and not in the final version
presented in this section, using the proper variable makes this
code easier to follow.