Optionally specify margin size #396
Whoever wants to shrink the margins, they can now do so via the new
optional command line arguments

--geometry:left
--geometry:right
--geometry:top
--geometry:bottom

which are default options in pandoc (see pandoc's documentation
[here](https://pandoc.org/demo/example33/6.2-variables.html)).

Their default value is ''. These options are not passed to pandoc when
their value is ''.

The argument parser added can be extended easily with other pandoc's
argument options. If the new options' default value is '' then they will
not be passed to pandoc.